[mandiant] Clean restructure mandiant dockerfile and module
