+++ external-import/mandiant/src/connector/__main__.py
@@ -1,13 +1,14 @@
-import sys
+import traceback
 import time
+import sys
 
-from mandiant.base import Mandiant
+from .base import Mandiant
 
 try:
     mandiantConnector = Mandiant()
     while True:
         mandiantConnector.run()
 except Exception as e:
-    print(e)
+    print(traceback.format_exc())
     time.sleep(10)
     sys.exit(0)
