+++ external-import/mandiant/src/connector/campaigns.py
