+++ external-import/mandiant/src/connector/constants.py
