+++ external-import/mandiant/src/connector/indicators.py
