+++ external-import/mandiant/src/connector/utils.py
